rename to TitleAndList for consistency & brevity
