+++ src/components/JoinUs/Learning.js
@@ -8,7 +8,7 @@ import { Subtitle } from '../Typography'
 
 import { Section, Separator, TitleAndBody } from './elements'
 import ExternalAnchor from '../Common/ExternalAnchor'
-import TitleAndList from '../Common/TitleAndListWithOptionalLogos'
+import TitleAndList from '../Common/TitleAndList'
 
 const InsightAnchor = styled(ExternalAnchor)`
   display: flex;
